# chore: Fair benchmarking of VectorizableUpdate.